requirements.txt
streamlit==1.28.1
streamlit-folium==0.13.0
langchain==0.1.0
langchain-google-genai==0.0.8
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
yfinance==0.2.32
pandas==2.1.3
pydantic==2.5.0
aiohttp==3.9.1
folium==0.14.0
